Publish portable version of the app
Publish a portable version of the app for use on Operating Systems (e.g. macOS) that don't easily like running a file downloaded from the internet in a ZIP file without lots of user-intervention.
Remove the macOS-specific version. Will consider putting it back if notarizing it with my Apple Developer Account is easy enough to do.